Extract DeprecatedAttributeResolver from NodeScopeResolver
