+++ src/Analyser/StmtHandler/FunctionHandler.php
@@ -7,6 +7,7 @@ use PhpParser\Node\Expr;
 use PhpParser\Node\Stmt;
 use PhpParser\Node\Stmt\Function_;
 use PhpParser\Node\Stmt\Return_;
+use PHPStan\Analyser\DeprecatedAttributeResolver;
 use PHPStan\Analyser\ExpressionResultStorage;
 use PHPStan\Analyser\GatheringNodeCallback;
 use PHPStan\Analyser\ImpurePoint;
@@ -35,6 +36,7 @@ final class FunctionHandler implements StmtHandler
 {
 
 	public function __construct(
+		private DeprecatedAttributeResolver $deprecatedAttributeResolver,
 		private PhpDocsResolver $phpDocsResolver,
 	)
 	{
@@ -66,7 +68,7 @@ final class FunctionHandler implements StmtHandler
 		}
 
 		if (!$isDeprecated) {
-			[$isDeprecated, $deprecatedDescription] = $nodeScopeResolver->getDeprecatedAttribute($scope, $stmt);
+			[$isDeprecated, $deprecatedDescription] = $this->deprecatedAttributeResolver->getDeprecatedAttribute($scope, $stmt);
 		}
 
 		$functionScope = $scope->enterFunction(
